Weight/app-weight/requirements.txt
flask
datetime
flask-mysql
cryptography
python-dotenv
werkzeug.utils
os